Fix: plugin refactored
