+++ GitHubRepos/.gitignore
@@ -14,3 +14,4 @@
 .cxx
 .DG_Store
 /buildSrc/build/
+/.idea/shelf/
