--- GitHubRepos/.idea/shelf/Uncommitted_changes_before_Checkout_at_26_3_22__9_10_PM__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_26_3_22,_9_10_PM_[Default_Changelist]" date="1648300213198" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_26_3_22,_9_10_PM_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 26/3/22, 9:10 PM [Default Changelist]" />
-</changelist>
