+++ GitHubRepos/buildSrc/build.gradle.kts
@@ -1,3 +1,4 @@
+
 plugins {
     `java-gradle-plugin`
     `kotlin-dsl`
@@ -15,6 +16,8 @@ object GlobalVersions {
     const val DAGGER = "2.41"
     const val KOTLIN = "1.6.10"
 }
+
+
 
 dependencies {
     implementation("com.google.devtools.ksp:symbol-processing-api:1.6.10-1.0.4")
