+++ GitHubRepos/gradle.properties
@@ -35,4 +35,6 @@ android.enableJetifier=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
 
+# KhalidAndroidPlugin properties
+enableBuildLogs = true
